+++ src/test/java/uk/co/reecedunn/intellij/plugin/xquery/tests/annotator/StringLiteralAnnotatorTest.kt
@@ -73,5 +73,25 @@ class StringLiteralAnnotatorTest : AnnotatorTestCase() {
         assertThat(annotations[1].textAttributes, `is`(SyntaxHighlighter.IDENTIFIER))
     }
 
+    fun testPITest_StringLiteral_NCName_InvalidCharacter() {
+        val file = parseText("processing-instruction(\"xml~stylesheet\")")
+        val annotations = annotateTree(file, StringLiteralAnnotator())
+        assertThat(annotations.size, `is`(2))
+
+        assertThat(annotations[0].severity, `is`(HighlightSeverity.INFORMATION))
+        assertThat(annotations[0].startOffset, `is`(24))
+        assertThat(annotations[0].endOffset, `is`(27))
+        assertThat(annotations[0].message, `is`(nullValue()))
+        assertThat(annotations[0].enforcedTextAttributes, `is`(TextAttributes.ERASE_MARKER))
+        assertThat(annotations[0].textAttributes, `is`(HighlighterColors.NO_HIGHLIGHTING))
+
+        assertThat(annotations[1].severity, `is`(HighlightSeverity.INFORMATION))
+        assertThat(annotations[1].startOffset, `is`(24))
+        assertThat(annotations[1].endOffset, `is`(27))
+        assertThat(annotations[1].message, `is`(nullValue()))
+        assertThat(annotations[1].enforcedTextAttributes, `is`(nullValue()))
+        assertThat(annotations[1].textAttributes, `is`(SyntaxHighlighter.IDENTIFIER))
+    }
+
     // endregion
 }